requirements.txt
spacy==3.7.4
pandas==2.3.1
python-dateutil==2.8.2
wikipedia-api==0.8.1
prettytable==3.16.0
